Fix exit on missing REPL history file
Fixes #3

So the problem seems to be that Rustyline returns an error if `Editor::load_history` is called with a file name that does not exist. Fix is just to ignore the error. The file will be created for the first time when `Editor::save_history` is called on closing the REPL.