Create an extern keyword for primitive definitions
This adds an `extern` expression that can be used to declare primitives. For example:

```pikelet
prim-u64-eq = extern "u64-eq" : U64 -> U64 -> Bool
```

This means that primitive definitions are declared in the source language itself, meaning that we could potentially add documentation to these definitions in the future. It also makes it much easier to add type annotations to these definitions because the types are asserted in the source code, rather than in the Rust implementation.