Avoid ambiguous syntax in function types
This switches to using `Fun` and `fun` in function types and values respecitvely, mirroring the syntax of record types and values. This avoides the following ambiguous syntax:

```
(x : A) -> B
```

This could either be parsed as:

- a dependent function type
- a non-dependent function with the input type annotated

I was attempting to get around this in [rust-nbe-for-mltt](https://github.com/brendanzab/rust-nbe-for-mltt) either by using parser combinators with ordered choice, or a hand-rolled recursive descent parser, but discovered that this would still require a significant amount of lookahead to resolve.

I’m still open to discussing this further though! The new syntax has the downside of being less succinct, and I still find capitalised keywords a tad odd.

Closes #110