House-cleaning of AST variants
This churns the history pretty badly, but I think it’s important. We group together common variants in a consistent order in their definitions and usages in the compiler passes. I’ve also tried to order from weaker to stronger claims, and so have put functions above records in the ASTs.