Allow dependencies between record term elements
This allows the elements of record terms to depend on each other. For
example, this expression is now valid:

```
record { x = 1, y = x }
    : Record { x : S32, y : S32 }
```

This has the helpful side-effect of fixing the scoping glued levels in
record terms. Before this change we would trigger a runtime panic on
rendering the error message for the following term:

```
record {
    Elem = String,
    elem = record {},
} : Record {
    Elem : Type,
    elem : Elem,
}
```

This is because `Elem` is not bound in the record term!
Now we render the following error:

```
error: mismatched types
  ┌─ examples/escaping-local.pi:3:12
  │
3 │     elem = record {},
  │            ^^^^^^^^^ expected `String`, found `Record {}`
```

This also paves the way for allowing us to elaborate let-expressions
into record terms in the core language, as is done in [1ML]. For
example:

```
let {
    x : S32 = 1,
    y : S32 = x * 2,
} in y - x
```

Could be sugar for:

```
(record { x = 1, y = x * 2, body = y - x }
    : Record { x : S32, y : S32, body : S32 })
        .body
```

[1ML]: http://mpi-sws.org/~rossberg/1ml/

This change also brings the typing rules of record terms closer to that of
function terms with labelled parameters, which I think is aesthetically
appealing, and could eventually help us bring functions and records much
closer together in terms of typing rules and syntax.

It does however mean that we no longer allow for out-of order elements
in record terms. For example the following is no longer valid:

```
record { y = 1, x = 2 }
    : Record { x : S32, y : S32 }
```

I'd like to eventually lift this restriction however, but it might
require us to do some  topological sorting of field elements in the
elaborator.