Avoid local gluing for now
This breaks on terms like:

```
((fun A a => a) : Fun (A : Type) -> A -> A) S32
```

With the error:

```
thread 'main' panicked at 'called `Option::unwrap()` on a `None` value', pikelet/src/lang/core/semantics.rs:490:52
```

This is because we are creating glued values for all variables. This is a problem when we do application for dependent functions, because the glued value no longer refers to valid binder.